+++ test/test_mixins/test_scripting.py
@@ -13,7 +13,7 @@ from fakeredis import _msgs as msgs
 from test import testtools
 from test.testtools import raw_command
 
-json_tests = pytest.importorskip("lupa")
+_ = pytest.importorskip("lupa")
 
 
 @pytest.mark.min_server("7")
@@ -523,12 +523,19 @@ def test_lua_log(r, caplog):
     script = r.register_script(script)
     with caplog.at_level("DEBUG"):
         script()
-    assert caplog.record_tuples == [
-        (logger.name, logging.DEBUG, "debug"),
-        (logger.name, logging.INFO, "verbose"),
-        (logger.name, logging.INFO, "notice"),
-        (logger.name, logging.WARNING, "warning"),
-    ]
+    assert (
+        len(
+            set(caplog.record_tuples).intersection(
+                {
+                    (logger.name, logging.DEBUG, "debug"),
+                    (logger.name, logging.INFO, "verbose"),
+                    (logger.name, logging.INFO, "notice"),
+                    (logger.name, logging.WARNING, "warning"),
+                }
+            )
+        )
+        == 4
+    )
 
 
 def test_lua_log_no_message(r: redis.Redis):
@@ -545,7 +552,7 @@ def test_lua_log_different_types(r, caplog):
     script = r.register_script(script)
     with caplog.at_level("DEBUG"):
         script()
-    assert caplog.record_tuples == [(logger.name, logging.DEBUG, "string 1 3.14 string")]
+    assert len(set(caplog.record_tuples).intersection({(logger.name, logging.DEBUG, "string 1 3.14 string")})) == 1
 
 
 def test_lua_log_wrong_level(r: redis.Redis):
@@ -565,7 +572,7 @@ def test_lua_log_defined_vars(r, caplog):
     script = r.register_script(script)
     with caplog.at_level("DEBUG"):
         script()
-    assert caplog.record_tuples == [(logger.name, logging.DEBUG, "string")]
+    assert len(set(caplog.record_tuples).intersection({(logger.name, logging.DEBUG, "string")})) == 1
 
 
 def test_hscan_cursors_are_bytes(r: redis.Redis):